+++ psl-core/src/main/java/org/linqs/psl/model/rule/logical/UnweightedGroundLogicalRule.java
@@ -42,7 +42,7 @@ public class UnweightedGroundLogicalRule extends AbstractGroundLogicalRule
     }
 
     @Override
-    public double getInfeasibility() {
+    public float getInfeasibility() {
         return dissatisfaction.getValue();
     }
