+++ psl-core/src/main/java/org/linqs/psl/reasoner/term/TermGenerator.java
@@ -24,5 +24,5 @@ public interface TermGenerator<T extends ReasonerTerm, V extends ReasonerLocalVa
      * Use the ground rules in |ruleStore| to generate optimization terms and populate |termStore|.
      * @return the number of terms added to the term store.
      */
-    public int generateTerms(GroundRuleStore ruleStore, TermStore<T, V> termStore);
+    public long generateTerms(GroundRuleStore ruleStore, TermStore<T, V> termStore);
 }
